victara: dts: Disable partial updates
There is a problem with layers not getting updated correctly.
The easiest way to see this problem is to change a toggle in
Settings and as the toggle is redrawn, a rectangle is redrawn
but only part of it picks up the changed background.  If you have
trouble seeing it, long press before releasing and it's more
obvious.

(Actually, "the easiest" may be misleading, that is the only
rendering issue that I've noticed or any user has reported
at this time)

Change-Id: If9da0f219729daadd221f286a4b92573e8fceb9d